Rename json_utils.h to json-utils.h, and move related implementations to the json-utils.cpp.
