+++ src/lib/misc.cpp
@@ -84,7 +84,7 @@ __RCSID("$NetBSD: misc.c,v 1.41 2012/03/05 02:20:18 christos Exp $");
 #include "crypto/mem.h"
 #include "utils.h"
 #include "str-utils.h"
-#include "json_utils.h"
+#include "json-utils.h"
 
 #ifdef _WIN32
 #include <windows.h>
@@ -92,62 +92,6 @@ __RCSID("$NetBSD: misc.c,v 1.41 2012/03/05 02:20:18 christos Exp $");
 #include <codecvt>
 #define vsnprintf _vsnprintf
 #endif
-
-/* Shortcut function to add field checking it for null to avoid allocation failure.
-   Please note that it deallocates val on failure. */
-bool
-obj_add_field_json(json_object *obj, const char *name, json_object *val)
-{
-    if (!val) {
-        return false;
-    }
-    // TODO: in JSON-C 0.13 json_object_object_add returns bool instead of void
-    json_object_object_add(obj, name, val);
-    if (!json_object_object_get_ex(obj, name, NULL)) {
-        json_object_put(val);
-        return false;
-    }
-
-    return true;
-}
-
-bool
-obj_add_hex_json(json_object *obj, const char *name, const uint8_t *val, size_t val_len)
-{
-    if (val_len > 1024 * 1024) {
-        RNP_LOG("too large json hex field: %zu", val_len);
-        val_len = 1024 * 1024;
-    }
-
-    char   smallbuf[64] = {0};
-    size_t hexlen = val_len * 2 + 1;
-
-    char *hexbuf = hexlen < sizeof(smallbuf) ? smallbuf : (char *) malloc(hexlen);
-    if (!hexbuf) {
-        return false;
-    }
-
-    bool res = rnp::hex_encode(val, val_len, hexbuf, hexlen, rnp::HEX_LOWERCASE) &&
-               obj_add_field_json(obj, name, json_object_new_string(hexbuf));
-
-    if (hexbuf != smallbuf) {
-        free(hexbuf);
-    }
-    return res;
-}
-
-bool
-array_add_element_json(json_object *obj, json_object *val)
-{
-    if (!val) {
-        return false;
-    }
-    if (json_object_array_add(obj, val)) {
-        json_object_put(val);
-        return false;
-    }
-    return true;
-}
 
 const char *
 id_str_pair::lookup(const id_str_pair pair[], int id, const char *notfound)
